+++ back/weverse/src/main/java/com/weverse/sb/chat/service/ChatServiceImpl.java
@@ -1,0 +1,103 @@
+package com.weverse.sb.chat.service;
+
+import java.util.ArrayList;
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import com.weverse.sb.media.dto.LiveChatMessageDTO;
+import com.weverse.sb.media.dto.MediaChatMessageDTO;
+import com.weverse.sb.media.entity.LiveChatMessage;
+import com.weverse.sb.media.entity.MediaChatMessage;
+import com.weverse.sb.media.entity.Streaming;
+import com.weverse.sb.media.entity.UploadedVideo;
+import com.weverse.sb.media.repository.LiveChatMessageRepository;
+import com.weverse.sb.media.repository.MediaChatMessageRepository;
+import com.weverse.sb.media.repository.StreamingRepository;
+import com.weverse.sb.media.repository.UploadedVideoRepository;
+import com.weverse.sb.user.entity.User;
+import com.weverse.sb.user.repository.UserRepository;
+
+@Service
+public class ChatServiceImpl implements ChatService {
+
+	@Autowired
+	private StreamingRepository streamingRepository;
+
+	@Autowired
+	private UploadedVideoRepository uploadedVideoRepository;
+
+	@Autowired
+	private LiveChatMessageRepository liveChatMessageRepository;
+
+	@Autowired
+	private MediaChatMessageRepository mediaChatMessageRepository;
+
+	@Autowired
+	private UserRepository userRepository;
+
+	@Override
+	public void inputChatMessage(LiveChatMessageDTO chatMessage) {
+
+		User user = userRepository.findById(chatMessage.getUserId())
+				.orElseThrow(() -> new IllegalArgumentException("Invalid group ID"));
+		Streaming streaming = streamingRepository.findById(chatMessage.getStreamingId())
+				.orElseThrow(() -> new IllegalArgumentException("Invalid streaming ID"));
+
+		LiveChatMessage message = LiveChatMessage.builder()
+				.sentAt(chatMessage.getSentAt())
+				.streaming(streaming)
+				.user(user)
+				.nickname(chatMessage.getNickname())
+				.content(chatMessage.getContent())
+				.build();
+
+		liveChatMessageRepository.save(message);
+
+	}
+
+	@Override
+	public void inputMediaChatMessage(MediaChatMessageDTO mediaMessage) {
+
+		User user = userRepository.findById(mediaMessage.getUserId())
+				.orElseThrow(() -> new IllegalArgumentException("Invalid group ID"));
+		UploadedVideo video = uploadedVideoRepository.findById(mediaMessage.getMediaId())
+				.orElseThrow(() -> new IllegalArgumentException("Invalid video ID"));
+
+		MediaChatMessage message = MediaChatMessage.builder()
+				.sentAt(mediaMessage.getSentAt())
+				.uploadedVideo(video)
+				.user(user)
+				.nickname(mediaMessage.getNickname())
+				.content(mediaMessage.getContent())
+				.build();
+
+		mediaChatMessageRepository.save(message);
+
+	}
+
+	@Override
+	public List<MediaChatMessageDTO> getMessageList(Long mediaId) {
+		List<MediaChatMessage> chatList = mediaChatMessageRepository.findByUploadedVideoId(mediaId);
+		List<MediaChatMessageDTO> dtoList = new ArrayList<>();
+
+		for (MediaChatMessage msg : chatList) {
+
+			MediaChatMessageDTO dto = MediaChatMessageDTO.builder()
+					.mediaId(mediaId)
+					.userId(msg.getUser().getUserId())
+					.nickname(msg.getNickname())
+					.content(msg.getContent())
+					.sentAt(msg.getSentAt())
+					.build();
+
+			dtoList.add(dto);
+		}
+
+		return dtoList;
+	}
+
+
+
+}
