+++ back/weverse/src/main/java/com/weverse/sb/media/entity/MediaCategory.java
@@ -1,0 +1,24 @@
+package com.weverse.sb.media.entity;
+
+import jakarta.persistence.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Builder
+@Data
+@Entity
+@Table(name = "media_category")
+public class MediaCategory {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "category_id")
+    private Long categoryId;
+
+    @Column(name = "title", length = 255)
+    private String title;
+}
